bot_server/requirements.txt
python-telegram-bot>=22.7
fastapi>=0.135
uvicorn>=0.42
apscheduler>=3.11
